+++ src/app/store/podcast.const.ts
@@ -1,0 +1,85 @@
+import { LoadingState } from './podcast-store.model';
+import * as moment from 'moment';
+
+export const PODCAST_INTIAL_STATE = {
+  data: {
+    author: {
+      name: {
+        label: '',
+      },
+      uri: {
+        label: '',
+      },
+    },
+    updated: {
+      label: '',
+    },
+    rights: {
+      label: '',
+    },
+    title: {
+      label: '',
+    },
+    icon: {
+      label: '',
+    },
+    id: {
+      label: '',
+    },
+    link: [],
+    entry: [],
+  },
+  currentPodCast: {
+    info: {
+      wrapperType: '',
+      kind: '',
+      artistId: 0,
+      collectionId: 0,
+      trackId: 0,
+      artistName: '',
+      collectionName: '',
+      trackName: '',
+      collectionCensoredName: '',
+      trackCensoredName: '',
+      artistViewUrl: '',
+      collectionViewUrl: '',
+      feedUrl: '',
+      trackViewUrl: '',
+      artworkUrl30: '',
+      artworkUrl60: '',
+      artworkUrl100: '',
+      artworkUrl600: '',
+      collectionPrice: 0,
+      trackPrice: 0,
+      collectionHdPrice: 0,
+      releaseDate: '',
+      collectionExplicitness: '',
+      trackExplicitness: '',
+      trackCount: 0,
+      trackTimeMillis: 0,
+      country: '',
+      currency: '',
+      primaryGenreName: '',
+      contentAdvisoryRating: '',
+      genreIds: [],
+      genres: [],
+    },
+    tracks: [],
+  },
+  callState: LoadingState.INIT,
+};
+
+export const COMPARE_DATES = (baseDate: string): boolean => {
+  const currentDate = moment(baseDate);
+  const dateToCompare = moment(baseDate).add(24, 'hours');
+  const timeDifference = dateToCompare.diff(currentDate);
+  /**
+   * console.log('DATES => ', {
+    currentDate,
+    dateToCompare,
+    timeDifference,
+    pass24Hours: timeDifference > 24 * 60 * 60 * 1000,
+  });
+   */
+  return timeDifference > 24 * 60 * 60 * 1000;
+};