+++ src/environments/environment.ts
@@ -1,5 +1,6 @@
 export const environment = {
   production: false,
+  version: '0.0.2',
   assetsUrl: '/assets',
   publishDate: '25/10/2023',
 };