+++ CHANGELOG.md
@@ -7,6 +7,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 [Glossary](#glossary)
 
+## [0.0.2] - 2023-10-25
+
+### Added
+
+- Session storage service
+- PodCast store
+- Recovery data
+
 ## [0.0.1] - 2023-10-25
 
 ### Added